+fetch in genre hook
