+++ src/hooks/useGenre.ts
@@ -1,5 +1,6 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { GenreData } from "../Pages/Movies";
+import { APIKEY } from "../Services/api-client";
 
 export interface ValueData {
     id: number;
@@ -10,11 +11,26 @@ export interface ValueData {
   interface Props{
     value: ValueData[];
     setValue: (value: ValueData[]) => void;
+    type: string;
   }
   
-  const useGenre = ({setValue, value}: Props) => {
+  const useGenre = ({setValue, value, type}: Props) => {
     const [genre, setGenre] = useState<GenreData[]>([]);
     const [page, setPage] = useState(1);
+
+    useEffect(() => {
+      const fetchGenre = async () => {
+        const data = await fetch(
+          `https://api.themoviedb.org/3/genre/${type}/list?api_key=${APIKEY}&language=en-US`
+        );
+        const { genres } = await data.json();
+        console.log(genres);
+        setGenre(genres);
+      };
+      fetchGenre();
+    }, [setGenre, type]);
+    
+    console.log(page);
 
     const CategoryAdd = (genres: ValueData) => {
       setValue([...value, genres]);
